==> Seoul_Big_Data_Campus/1.Preprocess/3_wifi_preprocess2.ipynb <==
import urllib
import json

지명 주소를 위경도로 변환

위경도를 바탕으로 행정동을 추출하는 과정입니다.

import pandas as pd

data = pd.read_csv('data/err_data.csv',encoding='CP949')

data = data['x']

data

key='key'

API key는 요금 문제로 지워두겠습니다. 재현되지 않는점 양해 부탁드립니다.

밑의 코드의 흐름을 따라가시면 이해할 수 있습니다.

a = pd.DataFrame(columns = ['주소'])
b = pd.DataFrame(columns = ['위경도'])
for i in range(len(data)):
    print(i)
    print(data[i])
    # 지명 주소를 위/경도 주소로,
    base_url ='https://maps.googleapis.com/maps/api/geocode/json?language=ko'+\
                '&address='
    # 부분 주소를 통해서 찾기
    address = urllib.parse.quote(data[i])
    addr = json.loads(urllib.request.urlopen(base_url+address+key).read().decode('utf-8'))
    
    # 위치 정보만 뽑기
    for j in addr['results']:
        latlon = (j['geometry']['location'])
    lat_list = latlon['lat'],latlon['lng']
    intList = list(map(str,lat_list))
    latlng = ','.join(intList)
    b.loc[i] = latlng
    # google api
    base_url ='https://maps.googleapis.com/maps/api/geocode/json?language=ko'+\
                '&location_type=ROOFTOP'+\
                '&latlng='
    
    addr = json.loads(urllib.request.urlopen(base_url+latlng+key).read().decode('utf-8'))
    
    # 주소 정보만 뽑기
    for k in addr['results']:
        formatted_addr = k['formatted_address']
    print(formatted_addr)
    a.loc[i] = formatted_addr

a['위경도']=b

a

a.to_csv("err_addr2.csv",header=True,index=True)
